backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
rapidfuzz==3.9.7
pywin32==311

